# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API MoliCasillas",
        "description": "Documentación de la API para el Sistema de Casilla Electrónica MoliCasillas",
        "contact": {
            "email": "admin@molicasillas.gob.pe"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://molicasillas.test",
            "description": "Servidor de API MoliCasillas"
        }
    ],
    "paths": {
...
        "/api/v1/terms/current": {
            "get": {
                "tags": [
                    "Términos"
                ],
                "summary": "Obtener términos y condiciones vigentes",
                "operationId": "dc9b4e52fc94d16e49112bd0140bd98a",
                "responses": {
                    "200": {
                        "description": "Versión de términos vigente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/VersionTerminos"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "No hay términos activos"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/terms/accept": {
            "post": {
                "tags": [
                    "Términos"
                ],
                "summary": "Aceptar términos y condiciones",
                "operationId": "b5ea0f744ee9cd60eca67decbe1f85fb",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "version_terminos_id"
                                ],
                                "properties": {
                                    "version_terminos_id": {
                                        "type": "string",
                                        "format": "uuid"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Términos aceptados correctamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "properties": {
                                                "mensaje": {
                                                    "type": "string",
                                                    "example": "Términos aceptados correctamente."
                                                },
                                                "aceptado_en": {
                                                    "type": "string",
                                                    "format": "date-time"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Error de validación"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
...
            "VersionTerminos": {
                "title": "VersionTerminos",
                "description": "Versión de términos y condiciones",
                "properties": {
                    "id": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "version": {
                        "type": "string"
                    },
                    "contenido_html": {
                        "type": "string"
                    },
                    "es_activo": {
                        "type": "boolean"
                    },
                    "publicado_en": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}